Refactor debtors to use generic note controller
