+++ src/main/java/uk/gov/companieshouse/api/accounts/model/rest/smallfull/notes/debtors/CurrentPeriod.java
@@ -1,14 +1,25 @@
-package uk.gov.companieshouse.api.accounts.model.rest.notes.debtors;
+package uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.debtors;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.hibernate.validator.constraints.Range;
+import uk.gov.companieshouse.api.accounts.validation.CharSetValid;
+import uk.gov.companieshouse.charset.CharSet;
+
+import javax.validation.constraints.Size;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class PreviousPeriod {
+public class CurrentPeriod {
 
+    private static final int MAX_FIELD_LENGTH = 20000;
+    private static final int MIN_FIELD_LENGTH = 1;
     private static final int MAX_RANGE = 99999999;
     private static final int MIN_RANGE = 0;
+
+    @Size(min = MIN_FIELD_LENGTH, max = MAX_FIELD_LENGTH, message = "invalid.input.length")
+    @CharSetValid(CharSet.CHARACTER_SET_3)
+    @JsonProperty("details")
+    private String details;
 
     @Range(min = MIN_RANGE, max = MAX_RANGE, message = "value.outside.range")
     @JsonProperty("greater_than_one_year")
@@ -28,6 +39,14 @@ public class PreviousPeriod {
     @Range(min = MIN_RANGE, max = MAX_RANGE, message = "value.outside.range")
     @JsonProperty("trade_debtors")
     private Long tradeDebtors;
+
+    public String getDetails() {
+        return details;
+    }
+
+    public void setDetails(String details) {
+        this.details = details;
+    }
 
     public Long getGreaterThanOneYear() {
         return greaterThanOneYear;
