--- src/main/java/uk/gov/companieshouse/api/accounts/repository/DebtorsRepository.java
@@ -1,9 +1,0 @@
-package uk.gov.companieshouse.api.accounts.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-import uk.gov.companieshouse.api.accounts.model.entity.notes.debtors.DebtorsEntity;
-
-@Repository
-public interface DebtorsRepository  extends MongoRepository<DebtorsEntity, String> {
-}
