+++ src/main/java/uk/gov/companieshouse/api/accounts/repository/smallfull/StocksRepository.java
@@ -1,8 +1,9 @@
-package uk.gov.companieshouse.api.accounts.repository;
+package uk.gov.companieshouse.api.accounts.repository.smallfull;
 
 import org.springframework.stereotype.Repository;
 import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
 import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.stocks.StocksEntity;
+import uk.gov.companieshouse.api.accounts.repository.AccountsNoteRepository;
 
 @Repository
 public interface StocksRepository extends AccountsNoteRepository<StocksEntity> {
