+++ src/main/java/uk/gov/companieshouse/api/accounts/transformer/DebtorsTransformer.java
@@ -2,16 +2,17 @@ package uk.gov.companieshouse.api.accounts.transformer;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Component;
-import uk.gov.companieshouse.api.accounts.model.entity.notes.debtors.CurrentPeriodEntity;
-import uk.gov.companieshouse.api.accounts.model.entity.notes.debtors.DebtorsDataEntity;
-import uk.gov.companieshouse.api.accounts.model.entity.notes.debtors.DebtorsEntity;
-import uk.gov.companieshouse.api.accounts.model.entity.notes.debtors.PreviousPeriodEntity;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.debtors.CurrentPeriod;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.debtors.Debtors;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.debtors.PreviousPeriod;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
+import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.debtors.CurrentPeriodEntity;
+import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.debtors.DebtorsDataEntity;
+import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.debtors.DebtorsEntity;
+import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.debtors.PreviousPeriodEntity;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.debtors.CurrentPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.debtors.Debtors;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.debtors.PreviousPeriod;
 
 @Component
-public class DebtorsTransformer implements GenericTransformer<Debtors, DebtorsEntity> {
+public class DebtorsTransformer implements NoteTransformer<Debtors, DebtorsEntity> {
 
     @Override
     public DebtorsEntity transform(Debtors rest) {
@@ -58,5 +59,11 @@ public class DebtorsTransformer implements GenericTransformer<Debtors, DebtorsEn
             debtors.setPreviousPeriod(previousPeriod);
         }
         return debtors;
+    }
+
+    @Override
+    public AccountingNoteType getAccountingNoteType() {
+
+        return AccountingNoteType.SMALL_FULL_DEBTORS;
     }
 }
